# eap/openapi.yaml
openapi: 3.0.0
info:
    version: '1.0'
    title: 'TasteBuds API'
    description: 'Web Resources Specification (A7) for TasteBuds'
servers:
    -   url: http://lbaw2135-prod.fe.up.pt
        description: Production server
...
paths:
...
    /member/{username}/recipes:
        get: 
            operationId: R1007
            summary: 'R1007: Member recipes'
            description: 'Provide member recipes page. Access: PUB'
            tags: 
                - 'M02: Recipes, categories, groups and members'

            parameters:
                -   in: path
                    name: username
                    schema:
                        type: string
                    required: true

            responses:    
                '200':
                    description: 'Ok. Show [UI09](http://lbaw2135-piu.lbaw-prod.fe.up.pt/pages/profile.php)'
...
    /recipe/{recipeId}:
        get:
            operationId: R1011
            summary: 'R1011: View recipes'
            description: 'Show an individual recipe page. Access: PUB'
            tags:
                - 'M02: Recipes, categories, groups and members'
            parameters:
                -   in: path
                    name: recipeId
                    schema:
                        type: integer
                    required: true
            responses:
                '200':
                    description: 'Ok. Show [UI22](http://lbaw2135-piu.lbaw-prod.fe.up.pt/pages/recipe.php)'
...
    /api/user/{username}/recipes:
        get:
            operationId: R2005
            summary: 'R2005: Member recipes'
            description: "Gets the member's recipes in JSON. Access: PUB"
            tags: 
                - 'M02: Recipes, categories, groups and members'
            parameters:
                -   in: path
                    name: username
                    schema:
                        type: string
                    required: true
            responses:
                '200':
                  description: 'OK'
                  content:
                      application/json:
                          schema:
                              type: array
                              items:
                                  type: object
                                  properties:
                                      id:
                                          type: string
                                      title:
                                          type: string
                                      description:
                                          type: string
                                      creationDate:
                                          type: string
                                          format: datetime
                                      rating:
                                          type: number
                                      numberOfReviews:
                                          type: integer
                                      category:
                                          type: string
                                      tags:
                                          type: array
                                          items:
                                              type: object
                                              properties:
                                                  id: 
                                                      type: integer
                                                  name:
                                                      type: string
                                      inFavourites:
                                          type: boolean
                              example:
                                  -   id: 1
                                      title: Classic Tiramisu
                                      description: Classic Italian dessert made with ladyfingers and mascarpone cheese.
                                      creationDate: 22-01-2021
                                      rating: 4.5
                                      numberOfReviews: 5
                                      category: Dessert
                                      tags: 
                                          -   id: 45
                                              name: "Vegetarian"
                                          -   id: 32
                                              name: "Zero sugar"
                                      inFavourites: false
                                        
                '400':
                    description: 'Invalid Request'
                '403':
                    description: "The authenticated user cannot access the requested information (the requested member is private and the authenticated user doesn't follow him)"
                '404':
                    description: 'The user does not exist'
...
    /api/recipe:
        post:
            operationId: R2101
            summary: 'R2101: Create recipe'
            description: 'Creates a new recipe. Access: MBR'
            tags: 
                - 'M02: Recipes, categories, groups and members'
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                title:
                                    type: string
                                mainCategory:
                                    type: integer
                                description:
                                    type: string
                                tags:
                                    type: array
                                    items:
                                        type: integer
                                difficulty:
                                    type: string
                                servings:
                                    type: integer
                                images:
                                    type: array
                                    items:
                                        type: string
                                        format: binary
                                visibility:
                                    type: boolean
                                ingredients:
                                    type: array
                                    items:
                                        type: object
                                        properties:
                                            quantity:
                                                type: number
                                            unit:
                                                type: integer
                                            name:
                                                type: string
                                        required:
                                            - quantity
                                            - unit
                                            - name
                                preparationTime:
                                    type: integer
                                cookingTime:
                                    type: integer
                                additionalTime:
                                    type: integer
                                steps:
                                    type: array
                                    items:
                                        type: object
                                        properties:
                                            name:
                                                type: string
                                            description:
                                                type: string
                                            images:
                                                type: array
                                                items:
                                                    type: string
                                                    format: binary
                                        required:
                                            - name
                                            - description
                            required:
                                - title
                                - mainCategory
                                - description
                                - tags
                                - difficulty
                                - servings
                                - images
                                - visibility
                                - ingredients
                                - preparationTime
                                - cookingTime
                                - steps
            responses:
                '200':
                    description: 'Success'
                '400':
                    description: 'Invalid Request'
...